2025 day 3 part 2

+++ advent-of-code_2025/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod input;
 
-pub fn concat_nums(nums: &[u8]) -> u8 {
-    nums.iter().fold(0, |acc, &n| acc * 10 + n)
+pub fn concat_nums(nums: &[u8]) -> u64 {
+    nums.iter().fold(0, |acc, &n| acc * 10 + n as u64)
 }
